requirements.txt
requests
django
djangorestframework
mongoengine
django-rest-framework-mongoengine
tensorflow
numpy
